bypass TChecker's garbage collector
This is possible thanks to ticktac-project/tchecker#20 and has at
least three benefits: (1) we do not use a separate thread to scan the
memory continuously, (2) we avoid races conditions when the gc is
collecting pools that are being destructed, and (3) we don't leak any
TChecker type in the public API of tcltl.

Point (2) is the most important as it would have been a problem with
the future Python bindings, where we do not have strong control over
the order of destructions, and cannot ask interactive users to stop
the GC whenever an exception is raisd.

* src/tcltl.hh, bin/main.cc: Remove all references to the GC.
* src/tcltl.cc: Adjust accordingly, and call the destruct_state()
method of the allocator.  We have to delay those destruct_state()
using a queue because when Spot releases the state, than TChecker's
internals may still hold them.